requirements.txt
requests>=2.31.0
yfinance>=0.2.18
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
eventlet>=0.33.0